Add offset parameter for HTTP GET
It is now possible to define offset and limit parameters
in URL. This allows pagination of search results.

Changed the limit to be done by Sqlite database instead
of own filtering after reading the results from database.

This change should also improve memory usage with large
content sets because the Sqlite query does not always
return all the results. It is likely also that from CPU
point of view this is more efficient that post filtering
the results.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>