Change database query for counting totals
The REST pagination required the total count of entries in
database when limit and offset is used. This was previously
implemented with the select. This was felt that this was not
good from code maintenance point of view. Because of this,
the count is now separate method that can be called with
the same parameters as select. It just returns the total
count of this instead of rows.

Only internal refactoring and no external changes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>