Change Storage content meta creation
The Storage() object now returns contents which contain
hash with data and meta. This was intended to remove the
exposure of the hash structure in multiple files.

There is still exposure to hash internals in:

  contents['data'] = Migrate.content(contents['data'], self.content_type)

Even though this version is better than before, there is
still room for refactoring.

No external changes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>