Add meta object to HTTP GET responses
The meta field is now included always in HTTP GET responses.
The purpose of the field is to use it in pagination. The field
offers count, limit, offset and total attributes. These can
be used for e.g. to get information if all the search results
were returned in the response and what were the limit and
offset for the query.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>