Add pagination links for GET responses with offset
If GET requests has 'offset' parameter, the pagination links
are set into GET response. The reason for this is that it
was considered that the links may not make sense if the
offset parameter is not in the query string. And setting
or adding this autonomously from the server was considered
to be misleading and error prone.

User must explicitly request the pagination links by using
the offset parameter.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>