Refactor Logger creation
The logger creation previously used direct access to object
variable. This is now mimicking logging.getLogger() with own
getter.

The Logger class was tried to be set as a logger class with
the 'setLoggerClass' and by inheriting logging.Logger, like
in /1/. This was not going as expected. Logging.getLogger()
is a function and overriding that and the whole big picture
how to do this was not clear and the effort failed.

/1/ https://stackoverflow.com/a/28050837

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>