Fix log printing in logger configuration
If log message length was configured over the security length
check, the related log may have not been printed. If the log
level was previously set so that the debug logs were not printed,
the log was not printed. The reason was that the logger was
updated only after this log and the log used debug level.

This works now so that if user sets the debug or very verbose
in Logger configuration, the log is printed. If no logs are
requested, there is no log in this case. The reason is that
the specific log was not considered security event.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>