Fix log message operation ID (OID) refreshing
Each log contains OID which identifies one operation handled
by the server. The operation ID was refreshed always and it
was used to measure time of also content runtime. The content
runtime is inside REST API operation so that OID was changed
within the same operation.

The first operation handled by the server shares that same
OID is the server startup. This was not yet figured out how
to change this also for the very first operation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>