Refactor Content to Collection and Resource (partial)
Major refactoring to transform Content to Collection which
contains a list of Resources. The intention is to convey
Collection everywhere. For example Snippet or Solution
content are storing Collections and database queries return
this.

This allows metadata attached to resources and it tries to
provide better abstraction for the content.

This is partial commit and things are broken. First test
is partially working.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>