Refactor content template generation behind Collection
The intention is to avoid usage of Resource outside of the
Collection. For this reason, the Collection now has an API
interface that allows dumping content template.

Removed also unnecessary dump_json from Resource. This was
not dumping to JSON but to dictionary. And this was similar
than the dump_dict in Resource.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>