Refactor location of Constant module
The Constant class was moved under config package to main
level. It was considered that is more logical location for
the constants that are shared between all packages.

The Constant class itself probably breaks Pyhonic coding
standards and it's usage should be avoided. There are few
'constants' defined since they are used everywhere and these
are hard to isolate to any other class.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>