Fix character encoding when data is received
The character encoding is now corrected when data is received
from user. The encoding is internally now always unicode string.

Previously the code logic to parse was not logical because
author did not really understand the unicode and all the
differences between Python 2 and 3. Now this is a bit more
clearer - even though author cannot still claim to understand
this.

[1] https://learn.rmotr.com/python/understanding-unicode-in-python/strings-and-unicode/encodings-and-utf-8
[2] http://farmdev.com/talks/unicode/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>