Edit logs and update logging rules
The logging rule now states that variables in the middle of
log string must be separated with colon and space before and
after the variable. This applies to all logs and returned
causes for example in RESTish API server.

The intention is to provide standard way to log variables
in order to be able to extract the variables within log
strings for possible post processing.

Corrected also typos and clarified logs by adding more
variables for debugging.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>