Fix Collection count
The internal metadata for collection that holds the total
count of resources in collection was not updated. This field
is not currently used so there was no consequences from this
fault.

It is not yet clear if the field is going to be used. It was
decided to leave it into the code at this point of time.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>