Fix Operation ID (OID) refresh for the first request
The OID was not updated for the first request and the server
startup was done with the same OID. Now the server startup,
Gunicorn server startup and each request have always unique
OID that can be used to identify the transaction.

Added also stdout prints to tell user if the server is running
and when the server is exited. These logs obey -q (quiet), -vv
(very verbose) and --debug options like all other logs.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>