Add TLS/SSL certificates for HTTPS
Server now supports HTTPS. Security hardening is made so that
only TLSv1.2 is supported. All SSL version are considered
unsecure or compromised. TLS v1.1 and v1.2 are both without
known security issues, but only v1.2 provides modern
cryptographic algorithms [1].

Only ECDHE-RSA-AES256-SHA and ECDHE-RSA-AES128-SHA ciphers are
allowed with the TLSv1.2. This set balances compatibility and
security. Ciphers can be tested for example with script [2].
Cipher list and mapping can be read from [3].

There is no concept of user being supported in the server.
Because of this, there is no authentication.

[1] https://github.com/ssllabs/research/wiki/SSL-and-TLS-Deployment-Best-Practices
[1] https://superuser.com/a/224263
[3] https://testssl.sh/openssl-rfc.mapping.html

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>