Edit log message to identify measured method
The timeit decorator did not log the timed method. This caused
problems to identify which method measurement was in the log
message. The decorated method name is now printed in the log.

There was no way to print other than the method name. The
inspect that is being used in the Cause() was not tried because
of perform issues and compatibility problems because the Inspect
causes e.g. with Jupyter.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>