Fix partially tests with Falcon 1.4.1
The test mock for os.path.isfile caused problems with the
latest Falcon 1.4.1. There was import of mimetypes which
used isfile which was globally mocked in Snippy tests.

This is initially fixed so that the isfile is now mocked
with with context and pytest fixtures instead of global
mock.

This fix is not complete.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>