Refactor tool data folder structure
For example the storage schema and server related data was
moved under the data folder. The intention was to cleanup
the folder and file structure. Previous for example the
server related were located under different subfolders when
they are now all located under the server folder.

The template files were also renamed to remove the 'template'
from the filename because that was considered redundant and
repeating the folder name.

The solution example was removed because it was considered
that it was not needed since there are examples in solution
defaults.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>