Fix inserting multiple references
Creating second reference failed. The problem was that the
previous implementation assumed always that the content data
is mandatory and unique. This caused problem with links
because the links were first thought to be added without
content data which was always empty. This caused situation
where all references were stored with the same data.

The design is now that internally references are managed the
same as other content like snippet and solutions. Internally
the tool always allocates the same content from link to data
field. For user this is not visible and references can be
added without the content data from user point of view.

This was considered better approach than adding multiple
conditions around the code to check if category is reference
or something else. This should create more maintainable code
that executes always the same. There is just the trick to
copy the links to data in one place.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>