Fix sorting of links in configuration base class
The base class that parses the configuration source options
cannot use content category for code logic. The reason is
that the content category defaults to 'snippet' when it is
left out. The content category may be read reliably only
for example when input file is read.

Removed the conditional sort for links in case of reference
content. The reference links should not be sorted. The reason
is that it is assumed that the first link in has more relevance
than the second if there are multiple links in one reference.

The conditional soft of reference list is in TODO list.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>