Change reference content links sorting and fix bugs
The reference content links are not sorted anymore. The
assumption is that in case reference content has multiple
links, the order matters. It is expected that the first link
can be more relevant than the following links.

The behavior can be compared to snippet and solution content
data. In case user wants to store snippet data, the order must
matter. The analogy can be used with links as well.

Fix export based on digest without defining content category

If used exports content that explicitly results one content
to be exported, the exported file name can be automatically
set. In this case, there is no need to define the content
category to get correct file. If user does not define the
exported file from command line or the content does not
have file name defined, the default export file name is
now selected automatically based on content category.

Fix exported content links field to use tuple

When content was exported to file, the content links were
accidentally exported in list form instead of tuple. It
is considered that the exported content should not be
altered and thus tuple is preferred in export.

Fix export template file name for references content

The reference content exported template was accidentally
in plural form when all other content templates are in
singular form. Also the reference template is now in
singular form.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>