Edit test case documentation
The test case documentation is now part of standard Python
function documentation string. This is considered better
than adding proprietary strings that are parsed to find out
the test case documentation.

Also the workflow tags are removed.

The intention is to update the test case parser to support
reading the test case documentation and specific workflow
command in each test. The intention is to provide searchable
test case summary to find out what is tested and what is not.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>