Fix CLI performance test timeout
In slow environments and with Python 2, the CLI performance
test was taking 12 seconds when the maximum time was 10
seconds.

Python 2 is slower than Python 3 because the Sqlite in case
of Python 2 does not support memory database and file DB is
used.

The timer was just increased to comply the slow environment.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>