Fix Python 3.4 Dockerfile based tests
The Dockerfiles for Python 3.4 tests did not compile anymore
because the README.rst was required by setup.py and it was
missing.

Also the Dockerfile build failed if the tests failed. Now the
Dockerfiles are made so that the Docker image build does not
fail if the tests fail. The reason is that in case of failures,
it is expected that the user wants to start the container and
see what went wrong.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>