Fix REST API resources request JSON validation
The resources collection REST API does not have to contain
data but it must contain the links. The JSON media validation
was updated to accept check that either data or links field
is included in the request.

The schema validation exception is minimized. The error log
message is different between Python 2.7, Python 3.4 and
later because of Unicode string and formatting differences.
Now the schema validation string is always minimized so
that extra line breaks and extra spaces are removed. This
should also help reading the text from the JSON response.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>