Fix cause message in case of missing mandatory field
If the content is empty, it matches to two error conditions.
One test verifies that the content is not a template and
another one checks if the data is empty. In case the content
is just a template without any modification, both of these
tests are true.

The problem comes with the solution content compared to
references and snippets. In case of solution, the content
itself is the whole data that is parsed. In case of snippets
and references, the template contains separate fields.

The cause text that complains about missing mandatory field
data or links has higher priority than the message that
complains about the template. The complaint about missing
mandatory data is considered more clear for the end user.
And in case of REST API, all the errors are sent in the
response so identification of the problem is easier.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>