Change database field runalias to name
The runalias was originally planned to run specific snippet
examples. It was then considered that this is a risk from
security point of view and maybe not a good idea. Then the
reference content was implemented. For http link references
a 'name' is associated with the link. The brief may not be
short enough to use with the link name. For this reason, a
more common name was set for the field.

The intention is also that the name can be used as a short
alias for the content that is easier to remember than brief.

This breaks the backward compatibility of reference yaml data.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>