Add uuid1 to database as one column
The intention is to be able to use uuid1 to guarantee
uniqueness if there are multiple front-end databases that
are merged to one. The uuid1 is used because it contains
the host name and it is assumed that the Snippy in server
mode would be run for example in different containers with
different hostname. This should help to merge the data into
one master database if needed.

The uuid1 is not currently visible in REST API or in the
export and import formats like YAML. The reason was that
this required test case updates which were not done as of
now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>