Fix setuptool install and  possible race conditions
The setup.py was using imports to load project metadata. This
is considered bad practise and may cause race conditions when
installer is run

The special conditions for failure were not met with current
code but the new implementation is more user friendly. The new
implementation does not fail if maintainer adds new includes
that may cause race conditions as described in Python packaging
guide [1].

[1] https://packaging.python.org/guides/single-sourcing-package-version/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>