Add method to remove ANSI codes from log message
Logger class now offers a method which allows user to
remove ANSI codes from message. The purpose is to avoid
adding global configuration flag or other workarounds for
the problem that __str__ is not allowing parameters that
could be used to define if the color codes are used.

There are few objects like Resource or Collection that
implement __str__ and they print the content with colors.
This causes problem when the resource is printed to
terminal log. In this case if ANSI codes are used, the
color setting is left active if the printed log message
is truncated. This is not nice from end user point of view.

In order to avoid running the ANSI code removal regexp for
all log messages, this is left for the caller for possible
performance effects.

This creates situation where the log message, especially
with the --debug option, is not the actual string that
was created because the invisible ANSI codes are removed.
There may be some cases that cannot be debugged because
of this. This is also used as a reasoning to leave this
functionality for the calling point.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>