Fix exception logging and database debugging
If there is a database integrity error caused from digest or
uuid, the are now more logs to try to identify the conflict.

There is also now debug method that prints that can be
easily added into code during development phase if needed.

Fixed also few log message to follow the logging rules
to use comma when printing variables. The purpose of this
rule is to allow post processing of text formatted logs
if needed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>