Fix loading content from JSON, YAML and text files
In case content loaded from file source files like YAML,
JSON or text did not contain specific key, there was an
exception. Now if the imported file does not contain
required key or keys, default value will be used.

This does not check if the mandatory content data or links
is sent. In case mandatory content is missing, the later
error checks will trigger a failure cause for user.

This is improvement also in case the database content
changes. For example now there was the uuid added and
without this fix, the missing field would cause an
exception. Now if the uuid field is missing, the uuid
is created and it can be considered as a seamless upgrade
from user point of view.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>