Fix test teardown to remove database content
The test case did not remove the content of the database in
case of Python 3. In case of Python 2, the whole database
file gets removed but in case of Python 3, the file is in
memory. It seems that the database does not get immediately
deleted when snippy.release() is called or it takes some time
before the delete is done.

Author is not fully confident on this analysis.

Also the json_load, json_dump, yaml_load and yaml dump were
refactored. The problem was that these fixtures masked the
mock of a file. Test cases were using syntax for example
with json_load that was asserting if a specific filename was
used when opening a file. This was confusing because the
correct name is like 'mock_file' which directly indicates
that the check is for file.

The test cases are following format where the file is
mocked with context manager in each test case. This is
considered more readable than hiding this behind other
fixtures.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>