Add UUID for migrated JSON and YAML files
The UUID is now exported and import with the JSON and YAML
files. The intention is that the UUID is allocated once and
it will follow the content for rest of the life cycle as
long as the content is not deleted.

The purpose of the UUID was allow a possibility to migrate
content to one centralized database from multiple instances
of a local database. The UUID is type of uuid1 which contains
the hostname and time stamp which allow uniqueness over
multiple installations.

This update modifies all the files. The update in Resource()
class related to this patch was in the previous commmit.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>