Edit reference performance results
The API and CLI performance results were updated.

API performance felt the same as before and the results seemed
to be along same jitter as usual when visually observed. Some
of the results were not changed because the results felt the
same as before.

CLI performance felt a bit more sluggish than before. Visual
estimation would be that the performance metrics degraded
around 3-5 % compared to last results.

It is bit strange that the CLI felt a bit more slower than
before because the recent changes should have impacted
equally for the server and CLI usage. So the reason why the
CLI felt a bit slower is not known.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>