Refactor UUID column location in database schema
The UUID was moved just before timestamps. The intention is
that the order is logical in sense that the static variables
are first and before the digest. This allows calculating the
message digest over sequential order of the table columns.

Currently the digest is not calculated over all the new
columns because that requires change in test cases. This is
done when the columns are stabilizing.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>