Fix code documentation strings
There was a conflicting method argument documentation string
between Google and NumPy formats. Because of this, the method
parameter list did not render correctly in the generated
document. There was a newline missing after each parameter
before this fix.

Fixed also package locations after removing the database
package and merging it with the storage. The old package
hierarchy was still used in the code documentation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>