Add source field for content
There is now source field in database and in all migrated
content and REST API server.

The purpose of this field is to be able to link to the source
where the content was imported. In future for example there
can be import from other snippet or cheat sheet services. In
these cases it is planned to indicate where the content was
imported.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>