Fix log message in exception from database insert
If there was an integrity exception from the database, the
inserted count in logs was incorrect. The exception was not
counted as a failure. This created log and code logic that
executed without being aware that database insert had failed.

Refactored also 'cause' variable that was behaving differently
than the Cause() to 'error'. The name 'cause' was considered
misleading.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>