Fix Pylint 2.0.0 warnings
Upgraded the Pylint 2.0.0 and corrected new warnings.

The correct format to ignore pylint and flake8 warnings seems
to be like: # noqa pylint: disable=undefined-variable. Previous
format: # noqa: F821 # pylint: disable=undefined-variable did
not work anymore with the Pylint 2.0.0.

It is strange that only this place was causing problems
because the same syntax was being used in 4 other places. Also
the other places were changed to use '# noqa pylint:' format.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>