Fix database schema fields source, name and uuid
New and modified fields were updated to database schema.
Uuid and source are new fields and the 'runalias' field
was renamed to 'name'.

The use case of UUID was to be able to merge multiple
databases to one database.

The runalias was first designed to automatically run
specific snippet but this was considered as a security
issue. The 'name' is planned to be used with references
which are practically browser bookmarks. For example
Chrome contains field 'name' that is shorter than a
'brief' to identify specific bookmark. This should
make it easier to import and export for example Chrome
bookmarks in the future.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>