Fix log message to follow logging rules
Rule 'Logs from exceptions are printed in INFO level all
other logs in DEBUG' was not followed when the exit cause
was printed into logs with -vv or --debug options. Now also
this log is printed in debug level.

This does not change functionality that is visible for user.
Only the logging level was changed if the logs are activated.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>