Fix REST API responses for References data field
The 'data' field is empty in case of references because it
has now meaning in case of references which are about links.
The REST API response had the 'data' field set to contain
empty string as a first element in the list. The data list
must be an empty list.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>