Edit setup.py and install requirements
The modules required for console installation are now in
top of the setup.py module in the same section where the
extra dependencies are defined.

It was considered misleading to have the extra dependencies
in head of the file when the actual install dependencies
were "hidden" later in the setup.py file

Added two classifiers to indicate that the tool can be also
used as a server. Information technology classifier was
added as intended audience. The classifiers are pretty
limited to describe the content of the tool.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>