Add support for PyPy
PyPy was added as a supported runtime environment in setup
classifiers. Pytest test suite with running the server and
manual performance testing with api_rest_client were done
to verify the functionality.

Testing was made with PyPy version 5.7.0 for Python 2.7.13.

Added also the CPython in classifiers since the CPython
seems to be the C implementation for Python itself [1].

Currently tests are not run with the PyPy runtime.

[1] https://stackoverflow.com/a/17130986

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>