Fix REST API references data attribute handling
The REST API server incorrect sent the 'data' attribute when
it was not requested. In case of resources, the data is always
an empty list because it does not have meaning in case of
references. The 'links' attribute is data part of reference.

If user selects returned fields and 'data' is not part of
the list, the data attribute is not sent anymore in case
of reference content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>