Fix setup.py and pyflakes vs. flake8 installation
The flake 3.5.0 version requires pyflakes between versions
1.5.0 and 1.6.0. This causes conflict when installing
latest pyflakes 2.0.0.

This was noticed when pipenv was tried. There the fix was to
remove the version pinning from the pyflakes so that the
installer can handle the correct dependency.

It seems that the pyflakes development had been stalled a
bit. But there was the latest version 2.0.0 so this module
was left into development tooling. The flake8 seems to be
a bit more activate than the pyflakes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>