Fix requirements.txt according to pipenv
The requirements now follow the pipenv settings. The pipenv
currently contains the dependencies for the cli and REST
server.

The setup.py does not yet contain the server dependencies.
The reason for this is that the setup.py is used to create
the package for PyPI and the server is not yet wanted to
be part of the default installation.

The requirements.txt is considered duplicate information
compared to setup.py. For now the requirements.txt is left
just in case if someone tries to use it. The documentation
is planned to be updated later on to clarify the supported
installation methods.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>