Fix exception after pip install
The tool threw exception when run after pip install.

The Config() init method did not set the cls.source which was
called from _storage_file. That is, the class global variable
was not set all! It was amazing how this went unnoticed
through tests.

The reason is that all test case mock _storage_file in order
to get the test database for Python 2. For Python 3 database
is in memory and there is no effect in tests.

This change seems to prevent running test suite for Pytihon 3
and Python 2 at the same time. This may have worked by accident
before. Each test case runs the 'delete_all_content' which
delete the database content regardless of Python 2 (file) or
Python 3 (memory) database content. This may have worked
because the Config.source was not defined to be global
previously immediately at the initialization phase.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>