Add option to operate content with uuid from CLI
It is now possible to operate content from command line
with -u|--uuid option. The option works in a same way
than the digest option.

The difference between these two identities is that the
uuid is static once it has been allocated but the message
digest changes according to the content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>