Edit Python module versions
Updated third party modules for development and live installs.

Pytest and pylint modules were not updated to latest versions
because pipenv complained incompatible versions. This is very
likely just faulty module version requirement that has not
been updated.

The setup.py, Pipenv and requirements.txt are kept in synchrony
based on what pipenv is able to install without warnings. This
is the main reason why the latest versions were not used.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>