Fix Travis CI
The logging_tree dependency is twice in the setup.py and it
was removed because of this in previous commit. How ever,
the Travis CI installs the 'test' target and it does not
use the 'dev' target. It was considered that because the
'dev' target install also the documentation dependencies,
the logging_tree is set to two sets as of now.

This is used in the code but it is behind debug() methods
which are never supposed to be called in live installations.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>