Add OAS3.0 definitions
The OAS2.0 is still considered a main reference. There is now
a problem with OAS2.0 that seems to be OK with OAS3.0. The
/{category} and /{keywords} are colliding. In reality the
/{category} is one of snippets, solutions or references
which don't collide with search keywords (unless user decides
to use any of the three categories as a search keyword).

It may be that this is not the best possible API design. As o
now, this is meditated how to do this.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>