Fix REST API resource request with multiple hits
If explicit resource was requested like:

  > GET /snippy/api/app/v1/uuid/1/brief
  > GET /snippy/api/app/v1/digest/1/brief
  > GET /snippy/api/app/v1/solution/1/brief

where the result was multiple resources, it resulted a scenario
that was considered as an error. In this case the first resource
in the collection was always used. For example if resource is
requested to get a specific field, it is not sure if the first
hit is the one that user actually wanted.

It was considered that if there are no resources or multiple
resources in these kind of cases, it is better to return an
error.

The JSON API v1.0 [1] defines that "null is only an appropriate
response when the requested URL is one that might correspond to
a single resource, but doesn't currently". This was read so that
in case of multiple hits from unique resource request, a null
should be returned with 200 OK. But it was considered better
to return error instead of 200 OK with data part set to null.

It is also considered that 404 Not Found in cases when there is
no matches or multiple matches is simpler than setting different
error codes. The other option for multiple matches could be 409
Conflict that is used in case of import and update operations that
require unique content to be operated. The failure reason can be
determined from the error title.

This case is different if uuid or digest is used in query parameter
like in case:

  > GET /snippy/api/app/v1/group/docker?scat=snippet&uuid=1

In the above case, the result can be multiple contents in
returned collection.

[1] http://jsonapi.org/format/1.0/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>