Add REST API validation for scat parameter
The scat parameter is now validated that the values match
to correct category names. There was accidental usage of
plural forms like 'snippets' that work from CLI with the
REST API. This caused no content to be found and the
resulting error was not considered descriptive.

The CLI supports plural forms from the content categories.
The REST API sets more strict limits for the category in
order to keep the implementation simple. For example if
plural forms would be supported, a question would be
for example if the returned JSON data.type field would
be in plural or singular format.

The main reason for CLI to support the plural and singular
format was that author never remembers which one it is.
Also from the CLI interface, the mapping is simple with
the used Python argparse module.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>