Change group to groups (breaking) and fix create solution
Field group is now groups. The behavior of the groups field
is similar than tags and links which allow storing multiple
items.

The intention of original group field was to provide a
single group that gathers large amount of content. This is
still the target. But the thought process was that why limit
options for end user. It should be possible to store multiple
categories if needed. This was the reason why the behaviour
was changed. In most of the cases it is still assumed that
the groups contain a single item.

This is a breaking change because the API and default YAML
templates were changed. The content stored in YAML templates
can be updated by manually changing the string context to
list context like:

  ...
  filename: ''
  groups: docker
  links:
  - https://docs.docker.com/engine/reference/commandline/build/
  ...

to

  ...
  filename: ''
  groups:
  - docker
  links:
  - https://docs.docker.com/engine/reference/commandline/build/
  ...

Fixed the solution creation. The solution creation from editor
is broken in v0.8.0. The reason is that there is a mismatch
between template, code and test cases. The DATE field was
previously removed from the solution template because it was
considered unnecessary since the content fields contain the
created and updated fields. Also comparing templates with
message digest was not possible with changing data. This is
now fixed so that the code and tests are updated to reflect
the removal of DATE field from solution template.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>