Edit TODO message to give reason for content lenth
The latest refactoring that changed the groups field to groups
caused JSON response length to be increased more than it should
have. The reason is that in these two cases, the --compact-json
flag was not used. This causes the response JSON to be sent as
pretty printed. When the string was changed to array in the JSON,
it created additional newlines that increased the pretty printed
result JSON length.

The related cases were test_api_search_reference_012 and
test_api_search_snippet_016.

The response and the header length are correct in tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>