Fix internal error handling
Internal error handing was improved for merging collection.
This change does not have external impacts. The reason is
that the code never calls merging a collection to collection
which would cause an error. How ever, in order to better
handle such cases that are not supported, better internal
error handling was implemented.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>