Fix searching all categories with option --all
The search all still did not work as it should. The category
--all must search the content at once from all categories and
produce a list of content that is sorted as requested. For
example the default sort is based on content brief field and
this must produce list of content from all categories sorted
as one.

There was still the problem that search was made from one
category at the time and the content sort was not working as
expected. The fault also generated own queries from all
categories instead of one query.

This also fixes a problem where the database search had
unnecessary 'all' defined as content group. There is no
value 'all' that would be stored into the database.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>