Fix operation ID refresh at Snippy server startup
The operation ID (OID) was refreshed so that the Gunicorn
server logs had different OID than rest of the Snippy server
startup logs. At some point of time this may have been the
intended behaviour. But now it was considered that it is
better to have the same OID for all the logs that are printed
at the server start. This allows searching the logs that are
printed during the server startup, including the third party
module logs.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>